fix: arduinoMegaGnssDataLogger
The sketch was not working properly. The problem was identified to be the Serial buffers getting full. The sketch has been optimized to minimize the probability of this to happen, and by increasing the SERIAL_RX_BUFFER_SIZE. This needs to be done manually, so a comment has been included in the sketch description.